+++ docs/source/conf.py
@@ -238,10 +238,14 @@ templates_path = [
 autosummary_filename_map = {
     "torch.nn.utils.prune.identity": "torch.nn.utils.prune.identity_function",
     "torch.nn.utils.prune.Identity": "torch.nn.utils.prune.Identity_class",
+    "torch.optim.adafactor.adafactor": "torch.optim.adafactor.adafactor_function",
+    "torch.optim.adafactor.Adafactor": "torch.optim.adafactor.Adafactor_class",
     "torch.optim.adamw.adamw": "torch.optim.adamw.adamw_function",
     "torch.optim.adamw.AdamW": "torch.optim.adamw.AdamW_class",
     "torch.optim.asgd.asgd": "torch.optim.asgd.asgd_function",
     "torch.optim.asgd.ASGD": "torch.optim.asgd.ASGD_class",
+    "torch.optim.muon.muon": "torch.optim.muon.muon_function",
+    "torch.optim.muon.Muon": "torch.optim.muon.Muon_class",
     "torch.optim.nadam.nadam": "torch.optim.nadam.nadam_function",
     "torch.optim.nadam.NAdam": "torch.optim.nadam.NAdam_class",
     "torch.optim.radam.radam": "torch.optim.radam.radam_function",
