+++ torch/optim/adadelta.py
@@ -10,6 +10,7 @@ from .optimizer import (
     _differentiable_doc,
     _disable_dynamo_if_unsupported,
     _foreach_doc,
+    _functional_api_doc,
     _get_capturable_supported_devices,
     _get_scalar_dtype,
     _maximize_doc,
@@ -428,11 +429,6 @@ def adadelta(
     weight_decay: float,
     maximize: bool,
 ) -> None:
-    r"""Functional API that performs Adadelta algorithm computation.
-
-    See :class:`~torch.optim.Adadelta` for details.
-    """
-
     # this check is slow during compilation, so we skip it
     # if it's strictly needed we can add this check back in dynamo
     if not torch.compiler.is_compiling() and not all(
@@ -471,3 +467,6 @@ def adadelta(
         capturable=capturable,
         has_complex=has_complex,
     )
+
+
+adadelta.__doc__ = _functional_api_doc.format(optimizer="Adadelta")
