+++ torch/optim/adafactor.py
@@ -1,0 +1,4 @@
+from torch.optim._adafactor import Adafactor, adafactor
+
+
+__all__ = ["Adafactor", "adafactor"]
