+++ torch/optim/adamw.py
@@ -5,6 +5,7 @@ from .optimizer import (
     _capturable_doc,
     _differentiable_doc,
     _foreach_doc,
+    _functional_api_doc,
     _fused_doc,
     _maximize_doc,
     _params_doc,
@@ -151,10 +152,6 @@ def adamw(
     eps: float,
     maximize: bool,
 ) -> None:
-    r"""Functional API that performs AdamW algorithm computation.
-
-    See :class:`~torch.optim.AdamW` for details.
-    """
     adam(
         params,
         grads,
@@ -178,3 +175,14 @@ def adamw(
         maximize=maximize,
         decoupled_weight_decay=True,
     )
+
+
+adamw.__doc__ = (
+    _functional_api_doc.format(optimizer="AdamW")
+    + r"""
+
+.. note::
+    With ``fused=True``, CUDA supports FP32 parameters and gradients with
+    BF16 moment buffers. See :ref:`functional-adamw-bf16-state` for an example.
+"""
+)
