+++ torch/optim/muon.py
@@ -1,0 +1,4 @@
+from torch.optim._muon import Muon, muon
+
+
+__all__ = ["Muon", "muon"]
