+++ torch/optim/radam.py
@@ -12,6 +12,7 @@ from .optimizer import (
     _differentiable_doc,
     _disable_dynamo_if_unsupported,
     _foreach_doc,
+    _functional_api_doc,
     _get_capturable_supported_devices,
     _get_scalar_dtype,
     _get_value,
@@ -585,10 +586,6 @@ def radam(
     weight_decay: float,
     eps: float,
 ) -> None:
-    r"""Functional API that performs RAdam algorithm computation.
-
-    See :class:`~torch.optim.RAdam` for details.
-    """
     if not all(isinstance(t, torch.Tensor) for t in state_steps):
         raise RuntimeError(
             "API has changed, `state_steps` argument must contain a list of singleton tensors"
@@ -624,3 +621,6 @@ def radam(
         capturable=capturable,
         has_complex=has_complex,
     )
+
+
+radam.__doc__ = _functional_api_doc.format(optimizer="RAdam")
