+++ torch/optim/rprop.py
@@ -12,6 +12,7 @@ from .optimizer import (
     _differentiable_doc,
     _disable_dynamo_if_unsupported,
     _foreach_doc,
+    _functional_api_doc,
     _get_capturable_supported_devices,
     _get_scalar_dtype,
     _maximize_doc,
@@ -429,10 +430,6 @@ def rprop(
     etaminus: float,
     etaplus: float,
 ) -> None:
-    r"""Functional API that performs rprop algorithm computation.
-
-    See :class:`~torch.optim.Rprop` for details.
-    """
     # this check is slow during compilation, so we skip it
     # if it's strictly needed we can add this check back in dynamo
     if not torch.compiler.is_compiling() and not all(
@@ -470,3 +467,6 @@ def rprop(
         differentiable=differentiable,
         has_complex=has_complex,
     )
+
+
+rprop.__doc__ = _functional_api_doc.format(optimizer="Rprop")
