+++ torch/optim/sgd.py
@@ -11,6 +11,7 @@ from .optimizer import (
     _device_dtype_check_for_fused,
     _differentiable_doc,
     _foreach_doc,
+    _functional_api_doc,
     _fused_doc,
     _maximize_doc,
     _params_doc,
@@ -268,10 +269,6 @@ def sgd(
     nesterov: bool,
     maximize: bool,
 ) -> None:
-    r"""Functional API that performs SGD algorithm computation.
-
-    See :class:`~torch.optim.SGD` for details.
-    """
     # Respect when the user inputs False/True for foreach or fused. We only want to change
     # the default when neither have been user-specified. Note that we default to foreach
     # and pass False to use_fused. This is not a mistake--we want to give the fused impl
@@ -317,6 +314,9 @@ def sgd(
         grad_scale=grad_scale,
         found_inf=found_inf,
     )
+
+
+sgd.__doc__ = _functional_api_doc.format(optimizer="SGD")
 
 
 def _single_tensor_sgd(
